requirements.txt
alembic==1.7.7
aniso8601==9.0.1
appdirs==1.4.3
appnope==0.1.0
APScheduler
apturl
asgiref
attrs
autopep8
backcall
bcrypt
blinker
Brlapi
cachelib
cachetools
certifi
cffi
chardet
charset-normalizer
click
colorama
colorhash
command-not-found
configparser
convertdate
cryptography
cssselect
cssutils
cupshelpers
cycler
dataclasses
dbus-python
decorator
defer
distlib
distro
distro-info
dnspython
docker
docker-compose
dockerpty
docopt==0.6.2
duplicity==0.8.21
email-validator==1.3.0
fasteners==0.14.1
filelock==3.9.0
Flask==1.1.2
Flask-Bcrypt==1.0.1
flask-change-password==0.0.7
Flask-Dance==3.1.0
Flask-DebugToolbar==0.13.1
Flask-JWT-Extended==3.25.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Maintenance==0.0.1
Flask-Migrate==3.1.0
Flask-MonitoringDashboard==3.1.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Session==0.3.2
Flask-SQLAlchemy==2.5.1
Flask-Toastr==0.5.7
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
fonttools==4.29.1
future==0.18.2
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
httplib2==0.20.2
humanize==4.0.0
idna==2.10
importlib-metadata==4.8.3
importlib-resources==5.4.0
imutils==0.5.4
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.0
jeepney==0.7.1
Jinja2==3.0.1
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.6
markup==0.2
MarkupSafe==2.0.1
matplotlib==3.5.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
nose==1.3.7
numpy==1.22.2
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
parso==0.7.0
pdf2image==1.16.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
premailer==3.10.0
prompt-toolkit==3.0.5
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.6.0
pycairo==1.20.1
pycodestyle==2.6.0
pycparser==2.20
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyMeeus==0.5.12
PyNaCl==1.5.0
pyparsing==3.0.7
pyRFC3339==1.1
pyrsistent==0.18.1
pytesseract==0.3.10
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-editor==1.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.0
requests-oauthlib==1.3.0
scipy==1.10.0
SecretStorage==3.3.1
six==1.14.0
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.36.8
SqlDB==2.0.13
sqlparse==0.3.0
ssh-import-id==5.11
systemd-python==234
texttable==1.6.4
toml==0.10.2
torch==1.13.0
torchvision==0.2.1
traitlets==5.8.1
trash-cli==0.17.1.14
typing_extensions==4.4.0
tzlocal==2.0.0
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.2
URLObject==2.4.3
usb-creator==0.3.7
virtualenv==20.0.20
wadllib==1.3.6
wcwidth==0.1.9
websocket-client==1.2.3
Werkzeug==0.16.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
zipp==3.1.0
zope.event==4.6
zope.interface==5.5.2
